chore: setup vite + react frontend project
Opted for Vite as it's more modern and performant compared to Webpack.
React due to requirement, but I've opted for the Typescript React for
better typing support and handle errors which can be caught at
compilation time instead of runtime. This is just a base setup with
nothing else included.