feat: add new fields to AcademicProgramme and replace most table
There was an error when I tried to migrate the database. This happen after I changed the relationship and create a new table called "Department". The relationship I changed was "AcademicProgramme belong to Faculty", now it is "AcademicProgramme belong to Department" and "Department belong to Faculties". After getting the error, I don't know what else to do except redo the table.